requirements_dev.txt
notebook == 7.0.6
pandas == 2.1.4
numpy == 1.25.2
scikit-learn == 1.3.2
ydata-profiling == 4.6.3
ipywidgets == 8.1.1
click == 8.1.7
python-dotenv == 1.0.0
dvclive == 3.5.1
dvc-s3 == 3.0.1
boto3 == 1.33.12
xgboost == 2.0.3
pytest == 7.4.4
ipykernel == 6.29.3